refactor(allocator): split `Arena` into multiple files (#21446)
Pure refactor. Split the huge file `arena.rs` into multiple files, each below 500 lines. No substantive changes, just moving code around.
