--- crates/oxc_allocator/src/arena.rs
@@ -1,2172 +1,0 @@
-//! Arena allocator.
-//!
-//! This module was originally copied from `bumpalo` at commit a47f6d6b7b5fee9c99a285f0de80257a0a982ef3
-//! (2 commits after 3.20.2 release). Changes have been made since.
-
-#![expect(clippy::inline_always, clippy::undocumented_unsafe_blocks)]
-
-use std::{
-    alloc::{Layout, alloc, dealloc},
-    cell::Cell,
-    cmp::Ordering,
-    hint::unreachable_unchecked,
-    iter,
-    marker::PhantomData,
-    mem,
-    ptr::{self, NonNull},
-    slice, str,
-};
-
-use allocator_api2::alloc::{AllocError, Allocator};
-
-use crate::bumpalo_alloc::Alloc as BumpaloAlloc;
-pub use crate::bumpalo_alloc::AllocErr;
-
-#[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
-use crate::tracking::AllocationStats;
-
-/// An arena to allocate into.
-///
-/// # No `Drop`s
-///
-/// Objects that are allocated will never have their [`Drop`] implementation
-/// called &mdash; unless you do it manually yourself. This makes it relatively
-/// easy to leak memory or other resources.
-///
-/// If you have a type which internally manages
-///
-/// * an allocation from the global heap (e.g. [`Vec<T>`]),
-/// * open file descriptors (e.g. [`std::fs::File`]), or
-/// * any other resource that must be cleaned up (e.g. an `mmap`)
-///
-/// and relies on its `Drop` implementation to clean up the internal resource,
-/// then if you allocate that type with an `Arena`, you need to find a new way to
-/// clean up after it yourself.
-///
-/// Potential solutions are:
-///
-/// * Using [`oxc_allocator::Box::new_in`] instead of [`Arena::alloc`], that
-///   will drop wrapped values similarly to [`std::boxed::Box`]. Note that this
-///   requires enabling the `"boxed"` Cargo feature for this crate. **This is
-///   often the easiest solution.**
-///
-/// * Calling [`drop_in_place`][drop_in_place] or using
-///   [`std::mem::ManuallyDrop`][manuallydrop] to manually drop these types.
-///
-/// * Using [`oxc_allocator::Vec`] instead of [`std::vec::Vec`].
-///
-/// * Avoiding allocating these problematic types within an `Arena`.
-///
-/// Note that not calling `Drop` is memory safe! Destructors are never
-/// guaranteed to run in Rust, you can't rely on them for enforcing memory
-/// safety.
-///
-/// [`Drop`]: https://doc.rust-lang.org/std/ops/trait.Drop.html
-/// [`Vec<T>`]: https://doc.rust-lang.org/std/vec/struct.Vec.html
-/// [`std::fs::File`]: https://doc.rust-lang.org/std/fs/struct.File.html
-/// [drop_in_place]: https://doc.rust-lang.org/std/ptr/fn.drop_in_place.html
-/// [manuallydrop]: https://doc.rust-lang.org/std/mem/struct.ManuallyDrop.html
-/// [`oxc_allocator::Vec`]: crate::Vec
-/// [`std::vec::Vec`]: https://doc.rust-lang.org/std/vec/struct.Vec.html
-/// [`oxc_allocator::Box::new_in`]: crate::Box::new_in
-/// [`std::boxed::Box`]: https://doc.rust-lang.org/std/boxed/struct.Box.html
-///
-/// # Example
-///
-/// ```
-/// # use oxc_allocator::arena::Arena;
-///
-/// // Create a new arena.
-/// let arena = Arena::new();
-///
-/// // Allocate values into the arena.
-/// let forty_two = arena.alloc(42);
-/// assert_eq!(*forty_two, 42);
-///
-/// // Mutable references are returned from allocation.
-/// let mut n = arena.alloc(123u64);
-/// *n = 456;
-/// ```
-///
-/// # Allocation Methods Come in Many Flavors
-///
-/// There are various allocation methods on `Arena`, the simplest being
-/// [`alloc`][Arena::alloc]. The others exist to satisfy some combination of
-/// fallible allocation and initialization. The allocation methods are
-/// summarized in the following table:
-///
-/// <table>
-///   <thead>
-///     <tr>
-///       <th></th>
-///       <th>Infallible Allocation</th>
-///       <th>Fallible Allocation</th>
-///     </tr>
-///   </thead>
-///     <tr>
-///       <th>By Value</th>
-///       <td><a href="#method.alloc"><code>alloc</code></a></td>
-///       <td><a href="#method.try_alloc"><code>try_alloc</code></a></td>
-///     </tr>
-///     <tr>
-///       <th>Infallible Initializer Function</th>
-///       <td><a href="#method.alloc_with"><code>alloc_with</code></a></td>
-///       <td><a href="#method.try_alloc_with"><code>try_alloc_with</code></a></td>
-///     </tr>
-///   <tbody>
-///   </tbody>
-/// </table>
-///
-/// ## Fallible Allocation: The `try_alloc_` Method Prefix
-///
-/// These allocation methods let you recover from out-of-memory (OOM)
-/// scenarios, rather than raising a panic on OOM.
-///
-/// ```
-/// # use oxc_allocator::arena::Arena;
-///
-/// let arena = Arena::new();
-///
-/// match arena.try_alloc(MyStruct {
-///     // ...
-/// }) {
-///     Ok(my_struct) => {
-///         // Allocation succeeded.
-///     }
-///     Err(e) => {
-///         // Out of memory.
-///     }
-/// }
-///
-/// struct MyStruct {
-///     // ...
-/// }
-/// ```
-///
-/// ## Initializer Functions: The `_with` Method Suffix
-///
-/// Calling one of the generic `…alloc(x)` methods is essentially equivalent to
-/// the matching [`…alloc_with(|| x)`](?search=alloc_with). However if you use
-/// `…alloc_with`, then the closure will not be invoked until after allocating
-/// space for storing `x` on the heap.
-///
-/// This can be useful in certain edge-cases related to compiler optimizations.
-/// When evaluating for example `arena.alloc(x)`, semantically `x` is first put
-/// on the stack and then moved onto the heap. In some cases, the compiler is
-/// able to optimize this into constructing `x` directly on the heap, however
-/// in many cases it does not.
-///
-/// The `…alloc_with` functions try to help the compiler be smarter. In most
-/// cases doing for example `arena.try_alloc_with(|| x)` on release mode will be
-/// enough to help the compiler realize that this optimization is valid and
-/// to construct `x` directly onto the heap.
-///
-/// ### Warning
-///
-/// These functions critically depend on compiler optimizations to achieve their
-/// desired effect. This means that it is not an effective tool when compiling
-/// without optimizations on.
-///
-/// Even when optimizations are on, these functions do not **guarantee** that
-/// the value is constructed on the heap. To the best of our knowledge no such
-/// guarantee can be made in stable Rust as of 1.54.
-#[derive(Debug)]
-pub struct Arena<const MIN_ALIGN: usize = 1> {
-    // The current chunk we are bump allocating within.
-    current_chunk_footer: Cell<NonNull<ChunkFooter>>,
-    // Whether this `Arena` is allowed to allocate additional chunks when the current one is full.
-    can_grow: bool,
-    /// Used to track number of allocations made in this `Arena` when `track_allocations` feature is enabled
-    #[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
-    pub(crate) stats: AllocationStats,
-}
-
-#[repr(C)]
-#[repr(align(16))]
-#[derive(Debug)]
-struct ChunkFooter {
-    // Pointer to the start of this chunk allocation. This footer is always at
-    // the end of the chunk.
-    data: NonNull<u8>,
-
-    // The layout of this chunk's allocation.
-    layout: Layout,
-
-    // Link to the previous chunk.
-    //
-    // Note that the last node in the `prev` linked list is the canonical empty
-    // chunk, whose `prev` link points to itself.
-    prev: Cell<NonNull<ChunkFooter>>,
-
-    // Bump allocation finger that is always in the range `self.data..=self`.
-    ptr: Cell<NonNull<u8>>,
-}
-
-/// A wrapper type for the canonical, statically allocated empty chunk.
-///
-/// For the canonical empty chunk to be `static`, its type must be `Sync`, which
-/// is the purpose of this wrapper type. This is safe because the empty chunk is
-/// immutable and never actually modified.
-#[repr(transparent)]
-struct EmptyChunkFooter(ChunkFooter);
-
-unsafe impl Sync for EmptyChunkFooter {}
-
-static EMPTY_CHUNK: EmptyChunkFooter = EmptyChunkFooter(ChunkFooter {
-    // This chunk is empty (except the foot itself).
-    layout: Layout::new::<ChunkFooter>(),
-
-    // The start of the (empty) allocatable region for this chunk is itself.
-    data: NonNull::from_ref(&EMPTY_CHUNK).cast::<u8>(),
-
-    // The end of the (empty) allocatable region for this chunk is also itself.
-    ptr: Cell::new(NonNull::from_ref(&EMPTY_CHUNK).cast::<u8>()),
-
-    // Invariant: the last chunk footer in all `ChunkFooter::prev` linked lists
-    // is the empty chunk footer, whose `prev` points to itself.
-    prev: Cell::new(NonNull::from_ref(&EMPTY_CHUNK.0)),
-});
-
-impl EmptyChunkFooter {
-    fn get(&'static self) -> NonNull<ChunkFooter> {
-        NonNull::from(&self.0)
-    }
-}
-
-impl ChunkFooter {
-    /// Returns the start and length of the currently allocated region of this chunk.
-    fn as_raw_parts(&self) -> (*mut u8, usize) {
-        let data = self.data.as_ptr().cast_const();
-        let ptr = self.ptr.get().as_ptr();
-        debug_assert!(data <= ptr);
-        debug_assert!(ptr.cast_const() <= ptr::from_ref::<ChunkFooter>(self).cast::<u8>());
-        // SAFETY: `ptr` is always before or equal to footer
-        let len =
-            unsafe { ptr::from_ref::<ChunkFooter>(self).cast::<u8>().offset_from_unsigned(ptr) };
-        (ptr, len)
-    }
-
-    /// Is this chunk the last empty chunk?
-    fn is_empty(&self) -> bool {
-        ptr::eq(self, EMPTY_CHUNK.get().as_ptr())
-    }
-}
-
-impl<const MIN_ALIGN: usize> Default for Arena<MIN_ALIGN> {
-    fn default() -> Self {
-        Self::with_min_align()
-    }
-}
-
-impl<const MIN_ALIGN: usize> Drop for Arena<MIN_ALIGN> {
-    fn drop(&mut self) {
-        unsafe {
-            dealloc_chunk_list(self.current_chunk_footer.get());
-        }
-    }
-}
-
-#[inline]
-unsafe fn dealloc_chunk_list(mut footer: NonNull<ChunkFooter>) {
-    unsafe {
-        while !footer.as_ref().is_empty() {
-            let f = footer;
-            footer = f.as_ref().prev.get();
-            dealloc(f.as_ref().data.as_ptr(), f.as_ref().layout);
-        }
-    }
-}
-
-// `Arena`s are safe to send between threads because nothing aliases its owned
-// chunks until you start allocating from it. But by the time you allocate from
-// it, the returned references to allocations borrow the `Arena` and therefore
-// prevent sending the `Arena` across threads until the borrows end.
-unsafe impl<const MIN_ALIGN: usize> Send for Arena<MIN_ALIGN> {}
-
-#[inline]
-fn is_pointer_aligned_to<T>(pointer: *mut T, align: usize) -> bool {
-    debug_assert!(align.is_power_of_two());
-
-    let pointer = pointer as usize;
-    let pointer_aligned = round_down_to(pointer, align);
-    pointer == pointer_aligned
-}
-
-#[inline]
-const fn round_up_to(n: usize, divisor: usize) -> Option<usize> {
-    debug_assert!(divisor > 0);
-    debug_assert!(divisor.is_power_of_two());
-    match n.checked_add(divisor - 1) {
-        Some(x) => Some(x & !(divisor - 1)),
-        None => None,
-    }
-}
-
-/// Like `round_up_to` but turns overflow into undefined behavior rather than
-/// returning `None`.
-#[inline]
-unsafe fn round_up_to_unchecked(n: usize, divisor: usize) -> usize {
-    if let Some(x) = round_up_to(n, divisor) {
-        x
-    } else {
-        debug_assert!(false, "round_up_to_unchecked failed");
-        unsafe { unreachable_unchecked() }
-    }
-}
-
-#[inline]
-fn round_down_to(n: usize, divisor: usize) -> usize {
-    debug_assert!(divisor > 0);
-    debug_assert!(divisor.is_power_of_two());
-    n & !(divisor - 1)
-}
-
-/// Same as `round_down_to` but preserves pointer provenance.
-#[inline]
-fn round_mut_ptr_down_to(ptr: *mut u8, divisor: usize) -> *mut u8 {
-    debug_assert!(divisor > 0);
-    debug_assert!(divisor.is_power_of_two());
-    ptr.wrapping_sub(ptr as usize & (divisor - 1))
-}
-
-#[inline]
-unsafe fn round_mut_ptr_up_to_unchecked(ptr: *mut u8, divisor: usize) -> *mut u8 {
-    debug_assert!(divisor > 0);
-    debug_assert!(divisor.is_power_of_two());
-    unsafe {
-        let aligned = round_up_to_unchecked(ptr as usize, divisor);
-        let delta = aligned - (ptr as usize);
-        ptr.add(delta)
-    }
-}
-
-// The typical page size these days.
-//
-// Note that we don't need to exactly match page size for correctness, and it is
-// okay if this is smaller than the real page size in practice. It isn't worth
-// the portability concerns and lack of const propagation that dynamically
-// looking up the actual page size implies.
-const TYPICAL_PAGE_SIZE: usize = 0x1000;
-
-// We only support alignments of up to 16 bytes for iter_allocated_chunks.
-const SUPPORTED_ITER_ALIGNMENT: usize = 16;
-pub const CHUNK_ALIGN: usize = SUPPORTED_ITER_ALIGNMENT;
-pub const CHUNK_FOOTER_SIZE: usize = mem::size_of::<ChunkFooter>();
-
-// Assert that `ChunkFooter` is at the supported alignment. This will give a
-// compile time error if it is not the case
-const _FOOTER_ALIGN_ASSERTION: () = {
-    assert!(mem::align_of::<ChunkFooter>() == CHUNK_ALIGN);
-};
-
-// Check the hard-coded value in `ast_tools` raw transfer generator is accurate.
-// We can only do this check if we're on a 64-bit little-endian platform with the `fixed_size` feature enabled,
-// because the `fixed_size_constants` module is only compiled under those conditions.
-// That's good enough, as the size of `ChunkFooter` only matters in that case anyway (Oxlint JS plugins).
-#[cfg(all(feature = "fixed_size", target_pointer_width = "64", target_endian = "little"))]
-const _: () = {
-    use crate::generated::fixed_size_constants::CHUNK_FOOTER_SIZE as EXPECTED_CHUNK_FOOTER_SIZE;
-    assert!(CHUNK_FOOTER_SIZE == EXPECTED_CHUNK_FOOTER_SIZE);
-};
-
-// Maximum typical overhead per allocation imposed by allocators.
-const MALLOC_OVERHEAD: usize = 16;
-
-// This is the overhead from malloc, footer and alignment. For instance, if
-// we want to request a chunk of memory that has at least X bytes usable for
-// allocations (where X is aligned to CHUNK_ALIGN), then we expect that the
-// after adding a footer, malloc overhead and alignment, the chunk of memory
-// the allocator actually sets aside for us is X+OVERHEAD rounded up to the
-// nearest suitable size boundary.
-const OVERHEAD: usize = match round_up_to(MALLOC_OVERHEAD + CHUNK_FOOTER_SIZE, CHUNK_ALIGN) {
-    Some(x) => x,
-    None => panic!(),
-};
-
-// The target size of our first allocation, including our overhead.
-// The available capacity will be slightly smaller.
-// 16 KiB covers the majority of real-world JS/TS files.
-const FIRST_ALLOCATION_GOAL: usize = 16 * 1024;
-
-// The actual size of the first allocation is going to be a bit smaller than the
-// goal. We need to make room for the footer, and we also need take the
-// alignment into account. We're trying to avoid this kind of situation:
-// https://blog.mozilla.org/nnethercote/2011/08/05/clownshoes-available-in-sizes-2101-and-up/
-const DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER: usize = FIRST_ALLOCATION_GOAL - OVERHEAD;
-
-/// The memory size and alignment details for a potential new chunk
-/// allocation.
-#[derive(Debug, Clone, Copy)]
-struct NewChunkMemoryDetails {
-    new_size_without_footer: usize,
-    align: usize,
-    size: usize,
-}
-
-/// Wrapper around `Layout::from_size_align` that adds debug assertions.
-#[inline]
-fn layout_from_size_align(size: usize, align: usize) -> Result<Layout, AllocErr> {
-    Layout::from_size_align(size, align).map_err(|_| AllocErr)
-}
-
-#[cold]
-#[inline(never)]
-fn allocation_size_overflow<T>() -> T {
-    panic!("requested allocation size overflowed")
-}
-
-// NB: We don't have constructors as methods on `impl<N> Arena<N>` that return
-// `Self` because then `rustc` can't infer the `N` if it isn't explicitly
-// provided, even though it has a default value. There doesn't seem to be a good
-// workaround, other than putting constructors on the `Arena<DEFAULT>`; even
-// `std` does this same thing with `HashMap`, for example.
-impl Arena<1> {
-    /// Construct a new arena to allocate into.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    /// let arena = Arena::new();
-    /// # let _ = arena;
-    /// ```
-    pub fn new() -> Self {
-        Self::with_capacity(0)
-    }
-
-    /// Attempt to construct a new arena to allocate into.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    /// let arena = Arena::try_new();
-    /// # let _ = arena.unwrap();
-    /// ```
-    #[expect(clippy::missing_errors_doc, reason = "`try_with_capacity(0)` always returns `Ok`")]
-    pub fn try_new() -> Result<Self, AllocErr> {
-        Arena::try_with_capacity(0)
-    }
-
-    /// Construct a new arena with the specified byte capacity to allocate into.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    /// let arena = Arena::with_capacity(100);
-    /// # let _ = arena;
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics if allocating the initial capacity fails.
-    pub fn with_capacity(capacity: usize) -> Self {
-        Self::try_with_capacity(capacity).unwrap_or_else(|_| oom())
-    }
-
-    /// Attempt to construct a new arena with the specified byte capacity to allocate into.
-    ///
-    /// Propagates errors when allocating the initial capacity.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::{AllocErr, Arena};
-    /// # fn _foo() -> Result<(), AllocErr> {
-    /// let arena = Arena::try_with_capacity(100)?;
-    /// # let _ = arena;
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err(AllocErr)` if any of:
-    ///
-    /// 1. A [`Layout`] cannot be constructed from `capacity` and `MIN_ALIGN`
-    ///    (for example because the rounded-up size overflows `isize`).
-    /// 2. Computing the new chunk's memory details overflows.
-    /// 3. The underlying global allocator fails to allocate the initial chunk.
-    ///
-    /// When `capacity` is `0` no allocation is performed and `Ok` is always returned.
-    pub fn try_with_capacity(capacity: usize) -> Result<Self, AllocErr> {
-        Self::try_with_min_align_and_capacity(capacity)
-    }
-}
-
-impl<const MIN_ALIGN: usize> Arena<MIN_ALIGN> {
-    /// Create a new `Arena` that enforces a minimum alignment.
-    ///
-    /// The minimum alignment must be a power of two and no larger than `16`.
-    ///
-    /// Enforcing a minimum alignment can speed up allocation of objects with
-    /// alignment less than or equal to the minimum alignment. This comes at the
-    /// cost of introducing otherwise-unnecessary padding between allocations of
-    /// objects with alignment less than the minimum.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// type ArenaAlign8 = Arena<8>;
-    /// let arena = ArenaAlign8::with_min_align();
-    /// for x in 0..u8::MAX {
-    ///     let x = arena.alloc(x);
-    ///     assert_eq!((x as *mut _ as usize) % 8, 0, "x is aligned to 8");
-    /// }
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics on invalid minimum alignments.
-    //
-    // Because of `rustc`'s poor type inference for default type/const
-    // parameters (see the comment above the `impl Arena` block with no const
-    // `MIN_ALIGN` parameter) and because we don't want to force everyone to
-    // specify a minimum alignment with `Arena::new()` et al, we have a separate
-    // constructor for specifying the minimum alignment.
-    pub fn with_min_align() -> Self {
-        assert!(MIN_ALIGN.is_power_of_two(), "MIN_ALIGN must be a power of two; found {MIN_ALIGN}");
-        assert!(
-            MIN_ALIGN <= CHUNK_ALIGN,
-            "MIN_ALIGN may not be larger than {CHUNK_ALIGN}; found {MIN_ALIGN}"
-        );
-
-        Self::new_impl(EMPTY_CHUNK.get())
-    }
-
-    /// Create a new `Arena` that enforces a minimum alignment and starts with
-    /// room for at least `capacity` bytes.
-    ///
-    /// The minimum alignment must be a power of two and no larger than `16`.
-    ///
-    /// Enforcing a minimum alignment can speed up allocation of objects with
-    /// alignment less than or equal to the minimum alignment. This comes at the
-    /// cost of introducing otherwise-unnecessary padding between allocations of
-    /// objects with alignment less than the minimum.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// type ArenaAlign8 = Arena<8>;
-    /// let mut arena = ArenaAlign8::with_min_align_and_capacity(8 * 100);
-    /// for x in 0..100_u64 {
-    ///     let x = arena.alloc(x);
-    ///     assert_eq!((x as *mut _ as usize) % 8, 0, "x is aligned to 8");
-    /// }
-    /// assert_eq!(
-    ///     arena.iter_allocated_chunks().count(), 1,
-    ///     "initial chunk had capacity for all allocations",
-    /// );
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics on invalid minimum alignments.
-    ///
-    /// Panics if allocating the initial capacity fails.
-    pub fn with_min_align_and_capacity(capacity: usize) -> Self {
-        Self::try_with_min_align_and_capacity(capacity).unwrap_or_else(|_| oom())
-    }
-
-    /// Create a new `Arena` that enforces a minimum alignment and starts with
-    /// room for at least `capacity` bytes.
-    ///
-    /// The minimum alignment must be a power of two and no larger than `16`.
-    ///
-    /// Enforcing a minimum alignment can speed up allocation of objects with
-    /// alignment less than or equal to the minimum alignment. This comes at the
-    /// cost of introducing otherwise-unnecessary padding between allocations of
-    /// objects with alignment less than the minimum.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::{AllocErr, Arena};
-    /// # fn _foo() -> Result<(), AllocErr> {
-    /// type ArenaAlign8 = Arena<8>;
-    /// let mut arena = ArenaAlign8::try_with_min_align_and_capacity(8 * 100)?;
-    /// for x in 0..100_u64 {
-    ///     let x = arena.alloc(x);
-    ///     assert_eq!((x as *mut _ as usize) % 8, 0, "x is aligned to 8");
-    /// }
-    /// assert_eq!(
-    ///     arena.iter_allocated_chunks().count(), 1,
-    ///     "initial chunk had capacity for all allocations",
-    /// );
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// Panics on invalid minimum alignments.
-    ///
-    /// Panics if allocating the initial capacity fails.
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err(AllocErr)` if any of:
-    ///
-    /// 1. A [`Layout`] cannot be constructed from `capacity` and `MIN_ALIGN`
-    ///    (for example because the rounded-up size overflows `isize`).
-    /// 2. Computing the new chunk's memory details overflows.
-    /// 3. The underlying global allocator fails to allocate the initial chunk.
-    ///
-    /// When `capacity` is `0` no allocation is performed and `Ok` is always returned.
-    pub fn try_with_min_align_and_capacity(capacity: usize) -> Result<Self, AllocErr> {
-        assert!(MIN_ALIGN.is_power_of_two(), "MIN_ALIGN must be a power of two; found {MIN_ALIGN}");
-        assert!(
-            MIN_ALIGN <= CHUNK_ALIGN,
-            "MIN_ALIGN may not be larger than {CHUNK_ALIGN}; found {MIN_ALIGN}"
-        );
-
-        if capacity == 0 {
-            return Ok(Self::new_impl(EMPTY_CHUNK.get()));
-        }
-
-        let layout = layout_from_size_align(capacity, MIN_ALIGN)?;
-
-        let chunk_footer = unsafe {
-            Self::new_chunk(
-                // `new_size_without_footer` here was `None` in original `bumpalo` code.
-                // Changed to `Some(capacity)` when we increased `FIRST_ALLOCATION_GOAL` to 16 KiB,
-                // to avoid `Arena::with_capacity` allocating 16 KiB even when requested `capacity` is much smaller.
-                Self::new_chunk_memory_details(Some(capacity), layout).ok_or(AllocErr)?,
-                layout,
-                EMPTY_CHUNK.get(),
-            )
-            .ok_or(AllocErr)?
-        };
-
-        Ok(Self::new_impl(chunk_footer))
-    }
-
-    /// Create a new `Arena` from a chunk footer pointer.
-    ///
-    /// This is a helper function for all code paths which create an `Arena`.
-    #[inline(always)]
-    fn new_impl(chunk_footer_ptr: NonNull<ChunkFooter>) -> Self {
-        Self {
-            current_chunk_footer: Cell::new(chunk_footer_ptr),
-            can_grow: true,
-            #[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
-            stats: AllocationStats::default(),
-        }
-    }
-
-    /// Get this arena's minimum alignment.
-    ///
-    /// All objects allocated in this arena get aligned to this value.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena2 = Arena::<2>::with_min_align();
-    /// assert_eq!(arena2.min_align(), 2);
-    ///
-    /// let arena4 = Arena::<4>::with_min_align();
-    /// assert_eq!(arena4.min_align(), 4);
-    /// ```
-    #[inline]
-    #[expect(clippy::unused_self, reason = "part of public API")]
-    pub fn min_align(&self) -> usize {
-        MIN_ALIGN
-    }
-
-    /// Determine the memory details including final size, alignment and final
-    /// size without footer for a new chunk that would be allocated to fulfill
-    /// an allocation request.
-    fn new_chunk_memory_details(
-        new_size_without_footer: Option<usize>,
-        requested_layout: Layout,
-    ) -> Option<NewChunkMemoryDetails> {
-        // We must have `CHUNK_ALIGN` or better alignment...
-        let align = CHUNK_ALIGN
-            // and we have to have at least our configured minimum alignment...
-            .max(MIN_ALIGN)
-            // and make sure we satisfy the requested allocation's alignment.
-            .max(requested_layout.align());
-
-        let mut new_size_without_footer =
-            new_size_without_footer.unwrap_or(DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER);
-
-        let requested_size =
-            round_up_to(requested_layout.size(), align).unwrap_or_else(allocation_size_overflow);
-        new_size_without_footer = new_size_without_footer.max(requested_size);
-
-        // We want our allocations to play nice with the memory allocator, and
-        // waste as little memory as possible. For small allocations, this means
-        // that the entire allocation including the chunk footer and mallocs
-        // internal overhead is as close to a power of two as we can go without
-        // going over. For larger allocations, we only need to get close to a
-        // page boundary without going over.
-        if new_size_without_footer < TYPICAL_PAGE_SIZE {
-            new_size_without_footer =
-                (new_size_without_footer + OVERHEAD).next_power_of_two() - OVERHEAD;
-        } else {
-            new_size_without_footer =
-                round_up_to(new_size_without_footer + OVERHEAD, TYPICAL_PAGE_SIZE)? - OVERHEAD;
-        }
-
-        debug_assert_eq!(align % CHUNK_ALIGN, 0);
-        debug_assert_eq!(new_size_without_footer % CHUNK_ALIGN, 0);
-        let size = new_size_without_footer
-            .checked_add(CHUNK_FOOTER_SIZE)
-            .unwrap_or_else(allocation_size_overflow);
-
-        Some(NewChunkMemoryDetails { new_size_without_footer, align, size })
-    }
-
-    /// Allocate a new chunk and return its initialized footer.
-    ///
-    /// If given, `layouts` is a tuple of the current chunk size and the
-    /// layout of the allocation request that triggered us to fall back to
-    /// allocating a new chunk of memory.
-    unsafe fn new_chunk(
-        new_chunk_memory_details: NewChunkMemoryDetails,
-        requested_layout: Layout,
-        prev: NonNull<ChunkFooter>,
-    ) -> Option<NonNull<ChunkFooter>> {
-        unsafe {
-            let NewChunkMemoryDetails { new_size_without_footer, align, size } =
-                new_chunk_memory_details;
-
-            let layout = layout_from_size_align(size, align).ok()?;
-
-            debug_assert!(size >= requested_layout.size());
-
-            let data = alloc(layout);
-            let data = NonNull::new(data)?;
-
-            // The `ChunkFooter` is at the end of the chunk.
-            let footer_ptr = data.as_ptr().add(new_size_without_footer);
-            debug_assert_eq!((data.as_ptr() as usize) % align, 0);
-            debug_assert_eq!(footer_ptr as usize % CHUNK_ALIGN, 0);
-            #[expect(
-                clippy::cast_ptr_alignment,
-                reason = "footer_ptr is aligned to CHUNK_ALIGN, which is == align_of::<ChunkFooter>()"
-            )]
-            let footer_ptr = footer_ptr.cast::<ChunkFooter>();
-
-            // The bump pointer is initialized to the end of the range we will bump
-            // out of, rounded down to the minimum alignment. It is the
-            // `NewChunkMemoryDetails` constructor's responsibility to ensure that
-            // even after this rounding we have enough non-zero capacity in the
-            // chunk.
-            let ptr = round_mut_ptr_down_to(footer_ptr.cast::<u8>(), MIN_ALIGN);
-            debug_assert_eq!(ptr as usize % MIN_ALIGN, 0);
-            debug_assert!(
-                data.as_ptr() <= ptr,
-                "bump pointer {ptr:#p} should still be greater than or equal to the \
-                 start of the bump chunk {data:#p}"
-            );
-            debug_assert_eq!((ptr as usize) - (data.as_ptr() as usize), new_size_without_footer);
-
-            let ptr = Cell::new(NonNull::new_unchecked(ptr));
-
-            ptr::write(footer_ptr, ChunkFooter { data, layout, prev: Cell::new(prev), ptr });
-
-            Some(NonNull::new_unchecked(footer_ptr))
-        }
-    }
-
-    /// Reset this arena.
-    ///
-    /// Performs mass deallocation on everything allocated in this arena by
-    /// resetting the pointer into the underlying chunk of memory to the start
-    /// of the chunk. Does not run any `Drop` implementations on deallocated
-    /// objects; see [the top-level documentation](struct.Arena.html) for details.
-    ///
-    /// If this arena has allocated multiple chunks to allocate into, then
-    /// the excess chunks are returned to the global allocator.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let mut arena = Arena::new();
-    ///
-    /// // Allocate a bunch of things.
-    /// {
-    ///     for i in 0..100 {
-    ///         arena.alloc(i);
-    ///     }
-    /// }
-    ///
-    /// // Reset the arena.
-    /// arena.reset();
-    ///
-    /// // Allocate some new things in the space previously occupied by the
-    /// // original things.
-    /// for j in 200..400 {
-    ///     arena.alloc(j);
-    /// }
-    ///```
-    pub fn reset(&mut self) {
-        #[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
-        self.stats.reset();
-
-        // Takes `&mut self` so `self` must be unique and there can't be any
-        // borrows active that would get invalidated by resetting.
-        unsafe {
-            if self.current_chunk_footer.get().as_ref().is_empty() {
-                return;
-            }
-
-            let cur_chunk = self.current_chunk_footer.get();
-
-            // Deallocate all chunks except the current one
-            let prev_chunk = cur_chunk.as_ref().prev.replace(EMPTY_CHUNK.get());
-            dealloc_chunk_list(prev_chunk);
-
-            // Reset the bump finger to the end of the chunk.
-            debug_assert!(
-                is_pointer_aligned_to(cur_chunk.as_ptr(), MIN_ALIGN),
-                "bump pointer {cur_chunk:#p} should be aligned to the minimum alignment of {MIN_ALIGN:#x}"
-            );
-            cur_chunk.as_ref().ptr.set(cur_chunk.cast());
-
-            debug_assert!(
-                self.current_chunk_footer.get().as_ref().prev.get().as_ref().is_empty(),
-                "We should only have a single chunk"
-            );
-            debug_assert_eq!(
-                self.current_chunk_footer.get().as_ref().ptr.get(),
-                self.current_chunk_footer.get().cast(),
-                "Our chunk's bump finger should be reset to the start of its allocation"
-            );
-        }
-    }
-
-    /// Allocate an object in this `Arena` and return an exclusive reference to
-    /// it.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for `T` fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.alloc("hello");
-    /// assert_eq!(*x, "hello");
-    /// ```
-    #[inline(always)]
-    pub fn alloc<T>(&self, val: T) -> &mut T {
-        self.alloc_with(|| val)
-    }
-
-    /// Try to allocate an object in this `Arena` and return an exclusive
-    /// reference to it.
-    ///
-    /// # Errors
-    ///
-    /// Errors if reserving space for `T` fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.try_alloc("hello");
-    /// assert_eq!(x, Ok(&mut "hello"));
-    /// ```
-    #[inline(always)]
-    pub fn try_alloc<T>(&self, val: T) -> Result<&mut T, AllocErr> {
-        self.try_alloc_with(|| val)
-    }
-
-    /// Pre-allocate space for an object in this `Arena`, initializes it using
-    /// the closure, then returns an exclusive reference to it.
-    ///
-    /// See [The `_with` Method Suffix](#initializer-functions-the-_with-method-suffix) for a
-    /// discussion on the differences between the `_with` suffixed methods and
-    /// those methods without it, their performance characteristics, and when
-    /// you might or might not choose a `_with` suffixed method.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for `T` fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.alloc_with(|| "hello");
-    /// assert_eq!(*x, "hello");
-    /// ```
-    #[expect(clippy::mut_from_ref)]
-    #[inline(always)]
-    pub fn alloc_with<F, T>(&self, f: F) -> &mut T
-    where
-        F: FnOnce() -> T,
-    {
-        #[inline(always)]
-        unsafe fn inner_writer<T, F>(ptr: *mut T, f: F)
-        where
-            F: FnOnce() -> T,
-        {
-            // This function is translated as:
-            // - allocate space for a T on the stack
-            // - call f() with the return value being put onto this stack space
-            // - memcpy from the stack to the heap
-            //
-            // Ideally we want LLVM to always realize that doing a stack
-            // allocation is unnecessary and optimize the code so it writes
-            // directly into the heap instead. It seems we get it to realize
-            // this most consistently if we put this critical line into it's
-            // own function instead of inlining it into the surrounding code.
-            unsafe { ptr::write(ptr, f()) };
-        }
-
-        let layout = Layout::new::<T>();
-
-        unsafe {
-            let p = self.alloc_layout(layout);
-            let p = p.as_ptr().cast::<T>();
-            inner_writer(p, f);
-            &mut *p
-        }
-    }
-
-    /// Tries to pre-allocate space for an object in this `Arena`, initializes
-    /// it using the closure, then returns an exclusive reference to it.
-    ///
-    /// See [The `_with` Method Suffix](#initializer-functions-the-_with-method-suffix) for a
-    /// discussion on the differences between the `_with` suffixed methods and
-    /// those methods without it, their performance characteristics, and when
-    /// you might or might not choose a `_with` suffixed method.
-    ///
-    /// # Errors
-    ///
-    /// Errors if reserving space for `T` fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.try_alloc_with(|| "hello");
-    /// assert_eq!(x, Ok(&mut "hello"));
-    /// ```
-    #[expect(clippy::mut_from_ref)]
-    #[inline(always)]
-    pub fn try_alloc_with<F, T>(&self, f: F) -> Result<&mut T, AllocErr>
-    where
-        F: FnOnce() -> T,
-    {
-        #[inline(always)]
-        unsafe fn inner_writer<T, F>(ptr: *mut T, f: F)
-        where
-            F: FnOnce() -> T,
-        {
-            // This function is translated as:
-            // - allocate space for a T on the stack
-            // - call f() with the return value being put onto this stack space
-            // - memcpy from the stack to the heap
-            //
-            // Ideally we want LLVM to always realize that doing a stack
-            // allocation is unnecessary and optimize the code so it writes
-            // directly into the heap instead. It seems we get it to realize
-            // this most consistently if we put this critical line into it's
-            // own function instead of inlining it into the surrounding code.
-            unsafe { ptr::write(ptr, f()) };
-        }
-
-        // Self-contained: `p` is allocated for `T` and then a `T` is written.
-        let layout = Layout::new::<T>();
-        let p = self.try_alloc_layout(layout)?;
-        let p = p.as_ptr().cast::<T>();
-
-        unsafe {
-            inner_writer(p, f);
-            Ok(&mut *p)
-        }
-    }
-
-    /// `Copy` a slice into this `Arena` and return an exclusive reference to
-    /// the copy.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.alloc_slice_copy(&[1, 2, 3]);
-    /// assert_eq!(x, &[1, 2, 3]);
-    /// ```
-    #[expect(clippy::mut_from_ref)]
-    #[inline(always)]
-    pub fn alloc_slice_copy<T>(&self, src: &[T]) -> &mut [T]
-    where
-        T: Copy,
-    {
-        let layout = Layout::for_value(src);
-        let dst = self.alloc_layout(layout).cast::<T>();
-
-        unsafe {
-            ptr::copy_nonoverlapping(src.as_ptr(), dst.as_ptr(), src.len());
-            slice::from_raw_parts_mut(dst.as_ptr(), src.len())
-        }
-    }
-
-    /// `Clone` a slice into this `Arena` and return an exclusive reference to
-    /// the clone. Prefer [`alloc_slice_copy`](#method.alloc_slice_copy) if `T` is `Copy`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// #[derive(Clone, Debug, Eq, PartialEq)]
-    /// struct Sheep {
-    ///     name: String,
-    /// }
-    ///
-    /// let originals = [
-    ///     Sheep { name: "Alice".into() },
-    ///     Sheep { name: "Bob".into() },
-    ///     Sheep { name: "Cathy".into() },
-    /// ];
-    ///
-    /// let arena = Arena::new();
-    /// let clones = arena.alloc_slice_clone(&originals);
-    /// assert_eq!(originals, clones);
-    /// ```
-    #[expect(clippy::mut_from_ref)]
-    #[inline(always)]
-    pub fn alloc_slice_clone<T>(&self, src: &[T]) -> &mut [T]
-    where
-        T: Clone,
-    {
-        let layout = Layout::for_value(src);
-        let dst = self.alloc_layout(layout).cast::<T>();
-
-        unsafe {
-            for (i, val) in src.iter().cloned().enumerate() {
-                ptr::write(dst.as_ptr().add(i), val);
-            }
-
-            slice::from_raw_parts_mut(dst.as_ptr(), src.len())
-        }
-    }
-
-    /// `Copy` a string slice into this `Arena` and return an exclusive reference to it.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the string fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let hello = arena.alloc_str("hello world");
-    /// assert_eq!("hello world", hello);
-    /// ```
-    #[expect(clippy::mut_from_ref)]
-    #[inline(always)]
-    pub fn alloc_str(&self, src: &str) -> &mut str {
-        let buffer = self.alloc_slice_copy(src.as_bytes());
-        unsafe {
-            // This is OK, because it already came in as str, so it is guaranteed to be utf8
-            str::from_utf8_unchecked_mut(buffer)
-        }
-    }
-
-    /// Allocates a new slice of size `len` into this `Arena` and returns an
-    /// exclusive reference to the copy.
-    ///
-    /// The elements of the slice are initialized using the supplied closure.
-    /// The closure argument is the position in the slice.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.alloc_slice_fill_with(5, |i| 5 * (i + 1));
-    /// assert_eq!(x, &[5, 10, 15, 20, 25]);
-    /// ```
-    #[expect(clippy::mut_from_ref)]
-    #[inline(always)]
-    pub fn alloc_slice_fill_with<T, F>(&self, len: usize, mut f: F) -> &mut [T]
-    where
-        F: FnMut(usize) -> T,
-    {
-        let layout = Layout::array::<T>(len).unwrap_or_else(|_| oom());
-        let dst = self.alloc_layout(layout).cast::<T>();
-
-        unsafe {
-            for i in 0..len {
-                ptr::write(dst.as_ptr().add(i), f(i));
-            }
-
-            let result = slice::from_raw_parts_mut(dst.as_ptr(), len);
-            debug_assert_eq!(Layout::for_value(result), layout);
-            result
-        }
-    }
-
-    /// Allocates a new slice of size `len` into this `Arena` and returns an
-    /// exclusive reference to the copy.
-    ///
-    /// All elements of the slice are initialized to `value`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.alloc_slice_fill_copy(5, 42);
-    /// assert_eq!(x, &[42, 42, 42, 42, 42]);
-    /// ```
-    #[inline(always)]
-    pub fn alloc_slice_fill_copy<T: Copy>(&self, len: usize, value: T) -> &mut [T] {
-        self.alloc_slice_fill_with(len, |_| value)
-    }
-
-    /// Allocates a new slice of size `len` slice into this `Arena` and return an
-    /// exclusive reference to the copy.
-    ///
-    /// All elements of the slice are initialized to `value.clone()`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let s: String = "Hello Arena!".to_string();
-    /// let x: &[String] = arena.alloc_slice_fill_clone(2, &s);
-    /// assert_eq!(x.len(), 2);
-    /// assert_eq!(&x[0], &s);
-    /// assert_eq!(&x[1], &s);
-    /// ```
-    #[inline(always)]
-    pub fn alloc_slice_fill_clone<T: Clone>(&self, len: usize, value: &T) -> &mut [T] {
-        self.alloc_slice_fill_with(len, |_| value.clone())
-    }
-
-    /// Allocates a new slice of size `len` slice into this `Arena` and return an
-    /// exclusive reference to the copy.
-    ///
-    /// The elements are initialized using the supplied iterator.
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails, or if the supplied
-    /// iterator returns fewer elements than it promised.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x: &[i32] = arena.alloc_slice_fill_iter([2, 3, 5].iter().cloned().map(|i| i * i));
-    /// assert_eq!(x, [4, 9, 25]);
-    /// ```
-    #[inline(always)]
-    pub fn alloc_slice_fill_iter<T, I>(&self, iter: I) -> &mut [T]
-    where
-        I: IntoIterator<Item = T>,
-        I::IntoIter: ExactSizeIterator,
-    {
-        let mut iter = iter.into_iter();
-        self.alloc_slice_fill_with(iter.len(), |_| {
-            iter.next().expect("Iterator supplied too few elements")
-        })
-    }
-
-    /// Allocates a new slice of size `len` slice into this `Arena` and return an
-    /// exclusive reference to the copy.
-    ///
-    /// All elements of the slice are initialized to [`T::default()`].
-    ///
-    /// [`T::default()`]: https://doc.rust-lang.org/std/default/trait.Default.html#tymethod.default
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space for the slice fails.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let x = arena.alloc_slice_fill_default::<u32>(5);
-    /// assert_eq!(x, &[0, 0, 0, 0, 0]);
-    /// ```
-    #[inline(always)]
-    pub fn alloc_slice_fill_default<T: Default>(&self, len: usize) -> &mut [T] {
-        self.alloc_slice_fill_with(len, |_| T::default())
-    }
-
-    /// Allocate space for an object with the given `Layout`.
-    ///
-    /// The returned pointer points at uninitialized memory, and should be
-    /// initialized with
-    /// [`std::ptr::write`](https://doc.rust-lang.org/std/ptr/fn.write.html).
-    ///
-    /// # Panics
-    ///
-    /// Panics if reserving space matching `layout` fails.
-    #[inline(always)]
-    pub fn alloc_layout(&self, layout: Layout) -> NonNull<u8> {
-        self.try_alloc_layout(layout).unwrap_or_else(|_| oom())
-    }
-
-    /// Attempts to allocate space for an object with the given `Layout` or else returns
-    /// an `Err`.
-    ///
-    /// The returned pointer points at uninitialized memory, and should be
-    /// initialized with
-    /// [`std::ptr::write`](https://doc.rust-lang.org/std/ptr/fn.write.html).
-    ///
-    /// # Errors
-    ///
-    /// Errors if reserving space matching `layout` fails.
-    #[inline(always)]
-    pub fn try_alloc_layout(&self, layout: Layout) -> Result<NonNull<u8>, AllocErr> {
-        let res = if let Some(p) = self.try_alloc_layout_fast(layout) {
-            Ok(p)
-        } else {
-            self.alloc_layout_slow(layout).ok_or(AllocErr)
-        };
-
-        #[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
-        if res.is_ok() {
-            self.stats.record_allocation();
-        }
-
-        res
-    }
-
-    #[inline(always)]
-    fn try_alloc_layout_fast(&self, layout: Layout) -> Option<NonNull<u8>> {
-        // We don't need to check for ZSTs here since they will automatically
-        // be handled properly: the pointer will be bumped by zero bytes,
-        // modulo alignment. This keeps the fast path optimized for non-ZSTs,
-        // which are much more common.
-        unsafe {
-            let footer_ptr = self.current_chunk_footer.get();
-            let footer = footer_ptr.as_ref();
-
-            let ptr = footer.ptr.get().as_ptr();
-            let start = footer.data.as_ptr();
-            debug_assert!(
-                start <= ptr,
-                "start pointer {start:#p} should be less than or equal to bump pointer {ptr:#p}"
-            );
-            debug_assert!(
-                ptr <= footer_ptr.cast::<u8>().as_ptr(),
-                "bump pointer {ptr:#p} should be less than or equal to footer pointer {footer_ptr:#p}"
-            );
-            debug_assert!(
-                is_pointer_aligned_to(ptr, MIN_ALIGN),
-                "bump pointer {ptr:#p} should be aligned to the minimum alignment of {MIN_ALIGN:#x}"
-            );
-            // This `match` should be boiled away by LLVM: `MIN_ALIGN` is a
-            // constant and the layout's alignment is also constant in practice
-            // after inlining.
-            let aligned_ptr = match layout.align().cmp(&MIN_ALIGN) {
-                Ordering::Less => {
-                    // We need to round the size up to a multiple of `MIN_ALIGN`
-                    // to preserve the minimum alignment. This might overflow
-                    // since we cannot rely on `Layout`'s guarantees.
-                    let aligned_size = round_up_to(layout.size(), MIN_ALIGN)?;
-
-                    let capacity = (ptr as usize) - (start as usize);
-                    if aligned_size > capacity {
-                        return None;
-                    }
-
-                    ptr.wrapping_sub(aligned_size)
-                }
-                Ordering::Equal => {
-                    // `Layout` guarantees that rounding the size up to its
-                    // align cannot overflow (but does not guarantee that the
-                    // size is initially a multiple of the alignment, which is
-                    // why we need to do this rounding).
-                    let aligned_size = round_up_to_unchecked(layout.size(), layout.align());
-
-                    let capacity = (ptr as usize) - (start as usize);
-                    if aligned_size > capacity {
-                        return None;
-                    }
-
-                    ptr.wrapping_sub(aligned_size)
-                }
-                Ordering::Greater => {
-                    // `Layout` guarantees that rounding the size up to its
-                    // align cannot overflow (but does not guarantee that the
-                    // size is initially a multiple of the alignment, which is
-                    // why we need to do this rounding).
-                    let aligned_size = round_up_to_unchecked(layout.size(), layout.align());
-
-                    let aligned_ptr = round_mut_ptr_down_to(ptr, layout.align());
-                    let capacity = (aligned_ptr as usize).wrapping_sub(start as usize);
-                    if aligned_ptr < start || aligned_size > capacity {
-                        return None;
-                    }
-
-                    aligned_ptr.wrapping_sub(aligned_size)
-                }
-            };
-
-            debug_assert!(
-                is_pointer_aligned_to(aligned_ptr, layout.align()),
-                "pointer {aligned_ptr:#p} should be aligned to layout alignment of {:#}",
-                layout.align()
-            );
-            debug_assert!(
-                is_pointer_aligned_to(aligned_ptr, MIN_ALIGN),
-                "pointer {aligned_ptr:#p} should be aligned to minimum alignment of {MIN_ALIGN:#}"
-            );
-            debug_assert!(
-                start <= aligned_ptr && aligned_ptr <= ptr,
-                "pointer {aligned_ptr:#p} should be in range {start:#p}..{ptr:#p}"
-            );
-
-            debug_assert!(!aligned_ptr.is_null());
-            let aligned_ptr = NonNull::new_unchecked(aligned_ptr);
-
-            footer.ptr.set(aligned_ptr);
-            Some(aligned_ptr)
-        }
-    }
-
-    /// Gets the remaining capacity in the current chunk (in bytes).
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::with_capacity(100);
-    ///
-    /// let capacity = arena.chunk_capacity();
-    /// assert!(capacity >= 100);
-    /// ```
-    pub fn chunk_capacity(&self) -> usize {
-        let current_footer = self.current_chunk_footer.get();
-        let current_footer = unsafe { current_footer.as_ref() };
-
-        current_footer.ptr.get().as_ptr() as usize - current_footer.data.as_ptr() as usize
-    }
-
-    /// Slow path allocation for when we need to allocate a new chunk,
-    /// because there isn't enough room in our current chunk.
-    #[inline(never)]
-    #[cold]
-    fn alloc_layout_slow(&self, layout: Layout) -> Option<NonNull<u8>> {
-        unsafe {
-            if !self.can_grow {
-                return None;
-            }
-
-            // Get a new chunk from the global allocator.
-            let current_footer = self.current_chunk_footer.get();
-            let current_layout = current_footer.as_ref().layout;
-
-            // By default, we want our new chunk to be about twice as big
-            // as the previous chunk. If the global allocator refuses it,
-            // we try to divide it by half until it works or the requested
-            // size is smaller than the default footer size.
-            let min_new_chunk_size = layout.size().max(DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER);
-            let mut base_size =
-                (current_layout.size() - CHUNK_FOOTER_SIZE).checked_mul(2)?.max(min_new_chunk_size);
-            let mut chunk_memory_details = iter::from_fn(|| {
-                if base_size >= min_new_chunk_size {
-                    let size = base_size;
-                    base_size /= 2;
-                    Self::new_chunk_memory_details(Some(size), layout)
-                } else {
-                    None
-                }
-            });
-
-            let new_footer = chunk_memory_details.find_map(|new_chunk_memory_details| {
-                Self::new_chunk(new_chunk_memory_details, layout, current_footer)
-            })?;
-
-            debug_assert_eq!(new_footer.as_ref().data.as_ptr() as usize % layout.align(), 0);
-
-            // Set the new chunk as our new current chunk.
-            self.current_chunk_footer.set(new_footer);
-
-            // And then we can rely on `try_alloc_layout_fast` to allocate
-            // space within this chunk.
-            let ptr = self.try_alloc_layout_fast(layout);
-            debug_assert!(ptr.is_some());
-            ptr
-        }
-    }
-
-    /// Returns an iterator over each chunk of allocated memory that
-    /// this arena has allocated into.
-    ///
-    /// The chunks are returned ordered by allocation time, with the most
-    /// recently allocated chunk being returned first, and the least recently
-    /// allocated chunk being returned last.
-    ///
-    /// The values inside each chunk are also ordered by allocation time, with
-    /// the most recent allocation being earlier in the slice, and the least
-    /// recent allocation being towards the end of the slice.
-    ///
-    /// # SAFETY
-    ///
-    /// Because this method takes `&mut self`, we know that the arena
-    /// reference is unique and therefore there aren't any active references to
-    /// any of the objects we've allocated in it either. This potential aliasing
-    /// of exclusive references is one common footgun for unsafe code that we
-    /// don't need to worry about here.
-    ///
-    /// However, there could be regions of uninitialized memory used as padding
-    /// between allocations, which is why this iterator has items of type
-    /// `[MaybeUninit<u8>]`, instead of simply `[u8]`.
-    ///
-    /// The only way to guarantee that there is no padding between allocations
-    /// or within allocated objects is if all of these properties hold:
-    ///
-    /// 1. Every object allocated in this arena has the same alignment,
-    ///    and that alignment is at most 16.
-    /// 2. Every object's size is a multiple of its alignment.
-    /// 3. None of the objects allocated in this arena contain any internal
-    ///    padding.
-    ///
-    /// If you want to use this `iter_allocated_chunks` method, it is *your*
-    /// responsibility to ensure that these properties hold before calling
-    /// `MaybeUninit::assume_init` or otherwise reading the returned values.
-    ///
-    /// Finally, you must also ensure that any values allocated into the arena
-    /// have not had their `Drop` implementations called on them.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let mut arena = Arena::new();
-    ///
-    /// // Allocate a bunch of `i32`s in this arena, potentially causing
-    /// // additional memory chunks to be reserved.
-    /// for i in 0..10000 {
-    ///     arena.alloc(i);
-    /// }
-    ///
-    /// // Iterate over each chunk we've allocated into. This is safe
-    /// // because we have only allocated `i32`s in this arena, which fulfills
-    /// // the above requirements.
-    /// for ch in arena.iter_allocated_chunks() {
-    ///     println!("Used a chunk that is {} bytes long", ch.len());
-    ///     println!("The first byte is {:?}", unsafe {
-    ///         ch[0].assume_init()
-    ///     });
-    /// }
-    ///
-    /// // Within a chunk, allocations are ordered from most recent to least
-    /// // recent. If we allocated 'a', then 'b', then 'c', when we iterate
-    /// // through the chunk's data, we get them in the order 'c', then 'b',
-    /// // then 'a'.
-    ///
-    /// arena.reset();
-    /// arena.alloc(b'a');
-    /// arena.alloc(b'b');
-    /// arena.alloc(b'c');
-    ///
-    /// assert_eq!(arena.iter_allocated_chunks().count(), 1);
-    /// let chunk = arena.iter_allocated_chunks().nth(0).unwrap();
-    /// assert_eq!(chunk.len(), 3);
-    ///
-    /// // Safe because we've only allocated `u8`s in this arena, which
-    /// // fulfills the above requirements.
-    /// unsafe {
-    ///     assert_eq!(chunk[0].assume_init(), b'c');
-    ///     assert_eq!(chunk[1].assume_init(), b'b');
-    ///     assert_eq!(chunk[2].assume_init(), b'a');
-    /// }
-    /// ```
-    pub fn iter_allocated_chunks(&mut self) -> ChunkIter<'_, MIN_ALIGN> {
-        // SAFETY: Ensured by mutable borrow of `self`.
-        let raw = unsafe { self.iter_allocated_chunks_raw() };
-        ChunkIter { raw, arena: PhantomData }
-    }
-
-    /// Returns an iterator over raw pointers to chunks of allocated memory that
-    /// this arena has allocated into.
-    ///
-    /// This is an unsafe version of [`iter_allocated_chunks()`](Arena::iter_allocated_chunks),
-    /// with the caller responsible for safe usage of the returned pointers as
-    /// well as ensuring that the iterator is not invalidated by new
-    /// allocations.
-    ///
-    /// # SAFETY
-    ///
-    /// Allocations from this arena must not be performed while the returned
-    /// iterator is alive. If reading the chunk data (or casting to a reference)
-    /// the caller must ensure that there exist no mutable references to
-    /// previously allocated data.
-    ///
-    /// In addition, all of the caveats when reading the chunk data from
-    /// [`iter_allocated_chunks()`](Arena::iter_allocated_chunks) still apply.
-    pub unsafe fn iter_allocated_chunks_raw(&self) -> ChunkRawIter<'_, MIN_ALIGN> {
-        ChunkRawIter { footer: self.current_chunk_footer.get(), arena: PhantomData }
-    }
-
-    /// Calculates the number of bytes currently allocated across all chunks in
-    /// this arena.
-    ///
-    /// If you allocate types of different alignments or types with
-    /// larger-than-typical alignment in the same arena, some padding
-    /// bytes might get allocated in the arena. Note that those padding
-    /// bytes will add to this method's resulting sum, so you cannot rely
-    /// on it only counting the sum of the sizes of the things
-    /// you've allocated in the arena.
-    ///
-    /// The allocated bytes do not include the size of arena metadata,
-    /// so the amount of memory requested from the Rust allocator is higher
-    /// than the returned value.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use oxc_allocator::arena::Arena;
-    ///
-    /// let arena = Arena::new();
-    /// let _x = arena.alloc_slice_fill_default::<u32>(5);
-    /// let bytes = arena.allocated_bytes();
-    /// assert!(bytes >= size_of::<u32>() * 5);
-    /// ```
-    pub fn allocated_bytes(&self) -> usize {
-        let mut total = 0;
-        let mut footer_ptr = self.current_chunk_footer.get();
-        // SAFETY: Walk the chunk list until the empty sentinel chunk.
-        // Every non-empty chunk is a live allocation whose `layout.size()` includes the footer.
-        unsafe {
-            while !footer_ptr.as_ref().is_empty() {
-                total += footer_ptr.as_ref().layout.size() - CHUNK_FOOTER_SIZE;
-                footer_ptr = footer_ptr.as_ref().prev.get();
-            }
-        }
-        total
-    }
-
-    #[inline]
-    unsafe fn is_last_allocation(&self, ptr: NonNull<u8>) -> bool {
-        let footer = self.current_chunk_footer.get();
-        let footer = unsafe { footer.as_ref() };
-        footer.ptr.get() == ptr
-    }
-
-    #[inline]
-    unsafe fn dealloc(&self, ptr: NonNull<u8>, layout: Layout) {
-        // If the pointer is the last allocation we made, we can reuse the bytes,
-        // otherwise they are simply leaked -- at least until somebody calls reset().
-        unsafe {
-            if self.is_last_allocation(ptr) {
-                let ptr = self.current_chunk_footer.get().as_ref().ptr.get();
-                let ptr = ptr.as_ptr().add(layout.size());
-
-                let ptr = round_mut_ptr_up_to_unchecked(ptr, MIN_ALIGN);
-                debug_assert!(
-                    is_pointer_aligned_to(ptr, MIN_ALIGN),
-                    "bump pointer {ptr:#p} should be aligned to the minimum alignment of {MIN_ALIGN:#x}"
-                );
-                let ptr = NonNull::new_unchecked(ptr);
-                self.current_chunk_footer.get().as_ref().ptr.set(ptr);
-            }
-        }
-    }
-
-    #[inline]
-    unsafe fn shrink(
-        &self,
-        ptr: NonNull<u8>,
-        old_layout: Layout,
-        new_layout: Layout,
-    ) -> Result<NonNull<u8>, AllocErr> {
-        // If the new layout demands greater alignment than the old layout has,
-        // then either
-        //
-        // 1. the pointer happens to satisfy the new layout's alignment, so we
-        //    got lucky and can return the pointer as-is, or
-        //
-        // 2. the pointer is not aligned to the new layout's demanded alignment,
-        //    and we are unlucky.
-        //
-        // In the case of (2), to successfully "shrink" the allocation, we have
-        // to allocate a whole new region for the new layout.
-        if old_layout.align() < new_layout.align() {
-            return if is_pointer_aligned_to(ptr.as_ptr(), new_layout.align()) {
-                Ok(ptr)
-            } else {
-                let new_ptr = self.try_alloc_layout(new_layout)?;
-
-                #[cfg(all(
-                    feature = "track_allocations",
-                    not(feature = "disable_track_allocations")
-                ))]
-                self.stats.record_reallocation_after_allocation();
-
-                // We know that these regions are nonoverlapping because
-                // `new_ptr` is a fresh allocation.
-                unsafe {
-                    ptr::copy_nonoverlapping(ptr.as_ptr(), new_ptr.as_ptr(), new_layout.size());
-                }
-
-                Ok(new_ptr)
-            };
-        }
-
-        debug_assert!(is_pointer_aligned_to(ptr.as_ptr(), new_layout.align()));
-
-        let old_size = old_layout.size();
-        let new_size = new_layout.size();
-
-        // This is how much space we would *actually* reclaim while satisfying
-        // the requested alignment.
-        let delta = round_down_to(old_size - new_size, new_layout.align().max(MIN_ALIGN));
-
-        if unsafe { self.is_last_allocation(ptr) }
-                // Only reclaim the excess space (which requires a copy) if it
-                // is worth it: we are actually going to recover "enough" space
-                // and we can do a non-overlapping copy.
-                //
-                // We do `old_size.div_ceil(2)` so division rounds up rather than
-                // down. Consider when:
-                //
-                //     old_size = 5
-                //     new_size = 3
-                //
-                // If we do not take care to round up, this will result in:
-                //
-                //     delta = 2
-                //     (old_size / 2) = (5 / 2) = 2
-                //
-                // And the the check will succeed even though we are have
-                // overlapping ranges:
-                //
-                //     |--------old-allocation-------|
-                //     |------from-------|
-                //                 |-------to--------|
-                //     +-----+-----+-----+-----+-----+
-                //     |  a  |  b  |  c  |  .  |  .  |
-                //     +-----+-----+-----+-----+-----+
-                //
-                // But we MUST NOT have overlapping ranges because we use
-                // `copy_nonoverlapping` below! Therefore, we round the division
-                // up to avoid this issue.
-                && delta >= old_size.div_ceil(2)
-        {
-            unsafe {
-                let footer = self.current_chunk_footer.get();
-                let footer = footer.as_ref();
-
-                // NB: new_ptr is aligned, because ptr *has to* be aligned, and we
-                // made sure delta is aligned.
-                let new_ptr = NonNull::new_unchecked(footer.ptr.get().as_ptr().add(delta));
-                debug_assert!(
-                    is_pointer_aligned_to(new_ptr.as_ptr(), MIN_ALIGN),
-                    "bump pointer {new_ptr:#p} should be aligned to the minimum alignment of {MIN_ALIGN:#x}"
-                );
-                footer.ptr.set(new_ptr);
-
-                // NB: we know it is non-overlapping because of the size check
-                // in the `if` condition.
-                ptr::copy_nonoverlapping(ptr.as_ptr(), new_ptr.as_ptr(), new_size);
-
-                #[cfg(all(
-                    feature = "track_allocations",
-                    not(feature = "disable_track_allocations")
-                ))]
-                self.stats.record_reallocation();
-
-                return Ok(new_ptr);
-            }
-        }
-
-        // If this wasn't the last allocation, or shrinking wasn't worth it,
-        // simply return the old pointer as-is.
-        Ok(ptr)
-    }
-
-    #[inline]
-    unsafe fn grow(
-        &self,
-        ptr: NonNull<u8>,
-        old_layout: Layout,
-        new_layout: Layout,
-    ) -> Result<NonNull<u8>, AllocErr> {
-        let old_size = old_layout.size();
-
-        let new_size = new_layout.size();
-        let new_size = round_up_to(new_size, MIN_ALIGN).ok_or(AllocErr)?;
-
-        let align_is_compatible = old_layout.align() >= new_layout.align();
-
-        if align_is_compatible && unsafe { self.is_last_allocation(ptr) } {
-            // Try to allocate the delta size within this same block so we can
-            // reuse the currently allocated space.
-            let delta = new_size - old_size;
-            if let Some(p) =
-                self.try_alloc_layout_fast(layout_from_size_align(delta, old_layout.align())?)
-            {
-                unsafe { ptr::copy(ptr.as_ptr(), p.as_ptr(), old_size) };
-
-                #[cfg(all(
-                    feature = "track_allocations",
-                    not(feature = "disable_track_allocations")
-                ))]
-                self.stats.record_reallocation();
-
-                return Ok(p);
-            }
-        }
-
-        // Fallback: do a fresh allocation and copy the existing data into it.
-        let new_ptr = self.try_alloc_layout(new_layout)?;
-        unsafe { ptr::copy_nonoverlapping(ptr.as_ptr(), new_ptr.as_ptr(), old_size) };
-
-        #[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
-        self.stats.record_reallocation_after_allocation();
-
-        Ok(new_ptr)
-    }
-}
-
-/// Methods only available when `from_raw_parts` feature is enabled.
-/// These methods are only used by raw transfer.
-#[cfg(feature = "from_raw_parts")]
-impl<const MIN_ALIGN: usize> Arena<MIN_ALIGN> {
-    /// Construct a static-sized [`Arena`] from an existing memory allocation.
-    ///
-    /// The [`Arena`] which is returned takes ownership of the memory allocation,
-    /// and the allocation will be freed if the `Arena` is dropped.
-    /// If caller wishes to prevent that happening, they must wrap the `Arena` in `ManuallyDrop`.
-    ///
-    /// The [`Arena`] returned by this function cannot grow.
-    ///
-    /// # SAFETY
-    ///
-    /// * `ptr` must be aligned on [`CHUNK_ALIGN`].
-    /// * `size` must be a multiple of [`CHUNK_ALIGN`].
-    /// * `size` must be at least [`CHUNK_FOOTER_SIZE`].
-    /// * The memory region starting at `ptr` and encompassing `size` bytes must be within a single allocation.
-    /// * The memory region starting at `ptr` and encompassing `size` bytes must have been allocated from system
-    ///   allocator with alignment of [`CHUNK_ALIGN`] (or caller must wrap the `Arena` in `ManuallyDrop`
-    ///   and ensure the backing memory is freed correctly themselves).
-    /// * `ptr` must have permission for writes.
-    ///
-    /// # Panics
-    ///
-    /// Panics on invalid minimum alignments.
-    pub unsafe fn from_raw_parts(ptr: NonNull<u8>, size: usize) -> Self {
-        // Validate `MIN_ALIGN`. These checks will be removed by compiler as `MIN_ALIGN` is statically known.
-        assert!(MIN_ALIGN.is_power_of_two(), "MIN_ALIGN must be a power of two; found {MIN_ALIGN}");
-        assert!(
-            MIN_ALIGN <= CHUNK_ALIGN,
-            "MIN_ALIGN may not be larger than {CHUNK_ALIGN}; found {MIN_ALIGN}"
-        );
-
-        // Debug assert that `ptr` and `size` fulfill size and alignment requirements
-        debug_assert!((ptr.as_ptr() as usize).is_multiple_of(CHUNK_ALIGN));
-        debug_assert!(size.is_multiple_of(CHUNK_ALIGN));
-        debug_assert!(size >= CHUNK_FOOTER_SIZE);
-
-        let size_without_footer = size - CHUNK_FOOTER_SIZE;
-
-        // Construct `ChunkFooter` and write into end of allocation.
-        // SAFETY: Caller guarantees:
-        // * `ptr` is the start of an allocation of `size` bytes.
-        // * `size` is `>= CHUNK_FOOTER_SIZE` - so `size - CHUNK_FOOTER_SIZE` cannot wrap around.
-        let chunk_footer_ptr = unsafe { ptr.add(size_without_footer) }.cast::<ChunkFooter>();
-        // SAFETY: Caller guarantees `size` is a multiple of `CHUNK_ALIGN`.
-        // Caller guarantees region from `ptr` to `ptr + size` forms a single allocation,
-        // so it must be a valid layout.
-        let layout = unsafe { Layout::from_size_align_unchecked(size, CHUNK_ALIGN) };
-        let chunk_footer = ChunkFooter {
-            data: ptr,
-            layout,
-            prev: Cell::new(EMPTY_CHUNK.get()),
-            ptr: Cell::new(chunk_footer_ptr.cast::<u8>()),
-        };
-
-        // SAFETY: If caller has upheld safety requirements, `chunk_footer_ptr` is `CHUNK_FOOTER_SIZE`
-        // bytes from the end of the allocation, and aligned on `CHUNK_ALIGN`.
-        // Therefore `chunk_footer_ptr` is valid for writing a `ChunkFooter`.
-        unsafe { chunk_footer_ptr.write(chunk_footer) };
-
-        // Create `Arena` and set `can_grow` to `false`.
-        // This means that the memory chunk we've just created will remain its only chunk.
-        // Therefore it can never be deallocated, until the `Arena` is dropped.
-        // `Arena::reset` would only reset the "cursor" pointer, not deallocate the memory.
-        let mut arena = Self::new_impl(chunk_footer_ptr);
-        arena.can_grow = false;
-        arena
-    }
-
-    /// Set cursor pointer for this [`Arena`]'s current chunk.
-    ///
-    /// This is dangerous, and this method should not ordinarily be used.
-    /// It is only here for manually resetting the `Arena`.
-    ///
-    /// # SAFETY
-    ///
-    /// * `Arena` must have at least 1 allocated chunk.
-    ///   It is UB to call this method on an `Arena` which has not allocated
-    ///   i.e. fresh from `Arena::new`.
-    /// * `ptr` must point to within the `Arena`'s current chunk.
-    /// * `ptr` must be equal to or after data pointer for this chunk.
-    /// * `ptr` must be equal to or before the chunk's `ChunkFooter`.
-    /// * `ptr` must be aligned to `MIN_ALIGN`.
-    /// * No live references to data in the current chunk before `ptr` can exist.
-    pub unsafe fn set_cursor_ptr(&self, ptr: NonNull<u8>) {
-        // SAFETY: `current_chunk_footer` always points to a valid `ChunkFooter`
-        let chunk_footer = unsafe { self.current_chunk_footer.get().as_ref() };
-
-        debug_assert!(ptr.as_ptr() >= chunk_footer.data.as_ptr());
-        debug_assert!(ptr.as_ptr().cast_const() <= ptr::from_ref(chunk_footer).cast::<u8>());
-        debug_assert!(ptr.addr().get().is_multiple_of(MIN_ALIGN));
-
-        // SAFETY: Caller guarantees `Arena` has at least 1 allocated chunk, and `ptr` is valid.
-        #[expect(clippy::unnecessary_safety_comment)]
-        chunk_footer.ptr.set(ptr);
-    }
-
-    /// Get pointer to end of the data region of this [`Arena`]'s current chunk
-    /// i.e to the start of the `ChunkFooter`.
-    pub fn data_end_ptr(&self) -> NonNull<u8> {
-        self.current_chunk_footer.get().cast::<u8>()
-    }
-
-    /// Get pointer to end of this [`Arena`]'s current chunk (after the `ChunkFooter`).
-    pub fn end_ptr(&self) -> NonNull<u8> {
-        let chunk_footer_ptr = self.current_chunk_footer.get();
-
-        // SAFETY: `chunk_footer_ptr` always points to a valid `ChunkFooter`,
-        // so stepping past it cannot be out of bounds of the chunk's allocation.
-        // If `Arena` has not allocated, so `chunk_footer_ptr` returns a pointer to the static empty chunk,
-        // it's still valid.
-        unsafe { chunk_footer_ptr.add(1).cast::<u8>() }
-    }
-}
-
-/// An iterator over each chunk of allocated memory that
-/// an arena has allocated into.
-///
-/// The chunks are returned ordered by allocation time, with the most recently
-/// allocated chunk being returned first.
-///
-/// The values inside each chunk are also ordered by allocation time, with the most
-/// recent allocation being earlier in the slice.
-///
-/// This struct is created by the [`iter_allocated_chunks`] method on
-/// [`Arena`]. See that function for a safety description regarding reading from the returned items.
-///
-/// [`Arena`]: struct.Arena.html
-/// [`iter_allocated_chunks`]: struct.Arena.html#method.iter_allocated_chunks
-#[derive(Debug)]
-pub struct ChunkIter<'a, const MIN_ALIGN: usize = 1> {
-    raw: ChunkRawIter<'a, MIN_ALIGN>,
-    arena: PhantomData<&'a mut Arena<MIN_ALIGN>>,
-}
-
-impl<'a, const MIN_ALIGN: usize> Iterator for ChunkIter<'a, MIN_ALIGN> {
-    type Item = &'a [mem::MaybeUninit<u8>];
-
-    fn next(&mut self) -> Option<Self::Item> {
-        unsafe {
-            let (ptr, len) = self.raw.next()?;
-            let slice = slice::from_raw_parts(ptr as *const mem::MaybeUninit<u8>, len);
-            Some(slice)
-        }
-    }
-}
-
-impl<const MIN_ALIGN: usize> iter::FusedIterator for ChunkIter<'_, MIN_ALIGN> {}
-
-/// An iterator over raw pointers to chunks of allocated memory that this
-/// arena has allocated into.
-///
-/// See [`ChunkIter`] for details regarding the returned chunks.
-///
-/// This struct is created by the [`iter_allocated_chunks_raw`] method on
-/// [`Arena`]. See that function for a safety description regarding reading from
-/// the returned items.
-///
-/// [`Arena`]: struct.Arena.html
-/// [`iter_allocated_chunks_raw`]: struct.Arena.html#method.iter_allocated_chunks_raw
-#[derive(Debug)]
-pub struct ChunkRawIter<'a, const MIN_ALIGN: usize = 1> {
-    footer: NonNull<ChunkFooter>,
-    arena: PhantomData<&'a Arena<MIN_ALIGN>>,
-}
-
-impl<const MIN_ALIGN: usize> Iterator for ChunkRawIter<'_, MIN_ALIGN> {
-    type Item = (*mut u8, usize);
-    fn next(&mut self) -> Option<(*mut u8, usize)> {
-        unsafe {
-            let foot = self.footer.as_ref();
-            if foot.is_empty() {
-                return None;
-            }
-            let (ptr, len) = foot.as_raw_parts();
-            self.footer = foot.prev.get();
-            Some((ptr, len))
-        }
-    }
-}
-
-impl<const MIN_ALIGN: usize> iter::FusedIterator for ChunkRawIter<'_, MIN_ALIGN> {}
-
-#[inline(never)]
-#[cold]
-fn oom() -> ! {
-    panic!("out of memory")
-}
-
-unsafe impl<const MIN_ALIGN: usize> BumpaloAlloc for &Arena<MIN_ALIGN> {
-    #[inline(always)]
-    unsafe fn alloc(&mut self, layout: Layout) -> Result<NonNull<u8>, AllocErr> {
-        self.try_alloc_layout(layout)
-    }
-
-    #[inline]
-    unsafe fn dealloc(&mut self, ptr: NonNull<u8>, layout: Layout) {
-        unsafe { Arena::<MIN_ALIGN>::dealloc(self, ptr, layout) };
-    }
-
-    #[inline]
-    unsafe fn realloc(
-        &mut self,
-        ptr: NonNull<u8>,
-        layout: Layout,
-        new_size: usize,
-    ) -> Result<NonNull<u8>, AllocErr> {
-        let old_size = layout.size();
-
-        if old_size == 0 {
-            return self.try_alloc_layout(layout);
-        }
-
-        let new_layout = layout_from_size_align(new_size, layout.align())?;
-        if new_size <= old_size {
-            unsafe { Arena::shrink(self, ptr, layout, new_layout) }
-        } else {
-            unsafe { Arena::grow(self, ptr, layout, new_layout) }
-        }
-    }
-}
-
-/// This function tests that Arena isn't Sync.
-/// ```compile_fail
-/// use oxc_allocator::arena::Arena;
-/// fn _requires_sync<T: Sync>(_value: T) {}
-/// fn _arena_not_sync(b: Arena) {
-///    _requires_sync(b);
-/// }
-/// ```
-#[cfg(doctest)]
-fn _doctest_only() {}
-
-unsafe impl<const MIN_ALIGN: usize> Allocator for &Arena<MIN_ALIGN> {
-    #[inline]
-    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
-        self.try_alloc_layout(layout)
-            .map(|p| unsafe {
-                NonNull::new_unchecked(ptr::slice_from_raw_parts_mut(p.as_ptr(), layout.size()))
-            })
-            .map_err(|_| AllocError)
-    }
-
-    #[inline]
-    unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout) {
-        unsafe { Arena::<MIN_ALIGN>::dealloc(self, ptr, layout) };
-    }
-
-    #[inline]
-    unsafe fn shrink(
-        &self,
-        ptr: NonNull<u8>,
-        old_layout: Layout,
-        new_layout: Layout,
-    ) -> Result<NonNull<[u8]>, AllocError> {
-        unsafe { Arena::<MIN_ALIGN>::shrink(self, ptr, old_layout, new_layout) }
-            .map(|p| unsafe {
-                NonNull::new_unchecked(ptr::slice_from_raw_parts_mut(p.as_ptr(), new_layout.size()))
-            })
-            .map_err(|_| AllocError)
-    }
-
-    #[inline]
-    unsafe fn grow(
-        &self,
-        ptr: NonNull<u8>,
-        old_layout: Layout,
-        new_layout: Layout,
-    ) -> Result<NonNull<[u8]>, AllocError> {
-        unsafe { Arena::<MIN_ALIGN>::grow(self, ptr, old_layout, new_layout) }
-            .map(|p| unsafe {
-                NonNull::new_unchecked(ptr::slice_from_raw_parts_mut(p.as_ptr(), new_layout.size()))
-            })
-            .map_err(|_| AllocError)
-    }
-
-    #[inline]
-    unsafe fn grow_zeroed(
-        &self,
-        ptr: NonNull<u8>,
-        old_layout: Layout,
-        new_layout: Layout,
-    ) -> Result<NonNull<[u8]>, AllocError> {
-        let mut ptr = unsafe { self.grow(ptr, old_layout, new_layout) }?;
-        (unsafe { ptr.as_mut() })[old_layout.size()..].fill(0);
-        Ok(ptr)
-    }
-}
-
-// NB: Only tests which require private types, fields, or methods should be in
-// here. Anything that can just be tested via public API surface should be in
-// `tests/arena/*`.
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // Uses private type `ChunkFooter`.
-    #[cfg(target_pointer_width = "64")]
-    #[test]
-    fn chunk_footer_is_six_words_on_64_bit() {
-        assert_eq!(mem::size_of::<ChunkFooter>(), mem::size_of::<usize>() * 6);
-    }
-
-    // Uses private type `ChunkFooter`.
-    #[cfg(target_pointer_width = "32")]
-    #[test]
-    fn chunk_footer_is_eight_words_on_32_bit() {
-        assert_eq!(mem::size_of::<ChunkFooter>(), mem::size_of::<usize>() * 8);
-    }
-
-    // Uses private `DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER`.
-    #[test]
-    fn allocated_bytes() {
-        let mut b = Arena::new();
-
-        assert_eq!(b.allocated_bytes(), 0);
-
-        b.alloc(0u8);
-
-        assert_eq!(b.allocated_bytes(), DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER);
-
-        b.reset();
-
-        assert_eq!(b.allocated_bytes(), DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER);
-    }
-
-    // Uses private `bumpalo_alloc` module.
-    #[test]
-    fn test_realloc() {
-        unsafe {
-            const CAPACITY: usize = DEFAULT_CHUNK_SIZE_WITHOUT_FOOTER;
-            let mut b = Arena::<1>::with_min_align_and_capacity(CAPACITY);
-
-            // `realloc` doesn't shrink allocations that aren't "worth it".
-            let layout = Layout::from_size_align(100, 1).unwrap();
-            let p = b.alloc_layout(layout);
-            let q = (&b).realloc(p, layout, 51).unwrap();
-            assert_eq!(p, q);
-            b.reset();
-
-            // `realloc` will shrink allocations that are "worth it".
-            let layout = Layout::from_size_align(100, 1).unwrap();
-            let p = b.alloc_layout(layout);
-            let q = (&b).realloc(p, layout, 50).unwrap();
-            assert!(p != q);
-            b.reset();
-
-            // `realloc` will reuse the last allocation when growing.
-            let layout = Layout::from_size_align(10, 1).unwrap();
-            let p = b.alloc_layout(layout);
-            let q = (&b).realloc(p, layout, 11).unwrap();
-            assert_eq!(q.as_ptr() as usize, p.as_ptr() as usize - 1);
-            b.reset();
-
-            // `realloc` will allocate a new chunk when growing the last
-            // allocation, if need be.
-            let layout = Layout::from_size_align(1, 1).unwrap();
-            let p = b.alloc_layout(layout);
-            let q = (&b).realloc(p, layout, CAPACITY + 1).unwrap();
-            assert_ne!(q.as_ptr() as usize, p.as_ptr() as usize - CAPACITY);
-            b.reset();
-
-            // `realloc` will allocate and copy when reallocating anything that
-            // wasn't the last allocation.
-            let layout = Layout::from_size_align(1, 1).unwrap();
-            let p = b.alloc_layout(layout);
-            let _ = b.alloc_layout(layout);
-            let q = (&b).realloc(p, layout, 2).unwrap();
-            assert!(q.as_ptr() as usize != p.as_ptr() as usize - 1);
-            b.reset();
-        }
-    }
-
-    // Uses our private `bumpalo_alloc` module.
-    #[test]
-    fn invalid_read() {
-        let mut b = &Arena::new();
-
-        unsafe {
-            let l1 = Layout::from_size_align(12000, 4).unwrap();
-            let p1 = BumpaloAlloc::alloc(&mut b, l1).unwrap();
-
-            let l2 = Layout::from_size_align(1000, 4).unwrap();
-            BumpaloAlloc::alloc(&mut b, l2).unwrap();
-
-            let p1 = b.realloc(p1, l1, 24000).unwrap();
-            let l3 = Layout::from_size_align(24000, 4).unwrap();
-            b.realloc(p1, l3, 48000).unwrap();
-        }
-    }
-}
