+++ crates/oxc_allocator/src/arena/drop.rs
@@ -1,0 +1,96 @@
+//! `Drop` implementation for `Arena`, and `reset` method.
+
+use std::{alloc, ptr::NonNull};
+
+use super::{Arena, ChunkFooter, EMPTY_CHUNK, utils::is_pointer_aligned_to};
+
+impl<const MIN_ALIGN: usize> Arena<MIN_ALIGN> {
+    /// Reset this arena.
+    ///
+    /// Performs mass deallocation on everything allocated in this arena by
+    /// resetting the pointer into the underlying chunk of memory to the start
+    /// of the chunk. Does not run any `Drop` implementations on deallocated
+    /// objects; see [the top-level documentation](struct.Arena.html) for details.
+    ///
+    /// If this arena has allocated multiple chunks to allocate into, then
+    /// the excess chunks are returned to the global allocator.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use oxc_allocator::arena::Arena;
+    ///
+    /// let mut arena = Arena::new();
+    ///
+    /// // Allocate a bunch of things.
+    /// {
+    ///     for i in 0..100 {
+    ///         arena.alloc(i);
+    ///     }
+    /// }
+    ///
+    /// // Reset the arena.
+    /// arena.reset();
+    ///
+    /// // Allocate some new things in the space previously occupied by the
+    /// // original things.
+    /// for j in 200..400 {
+    ///     arena.alloc(j);
+    /// }
+    ///```
+    pub fn reset(&mut self) {
+        #[cfg(all(feature = "track_allocations", not(feature = "disable_track_allocations")))]
+        self.stats.reset();
+
+        // Takes `&mut self` so `self` must be unique and there can't be any
+        // borrows active that would get invalidated by resetting.
+        unsafe {
+            if self.current_chunk_footer.get().as_ref().is_empty() {
+                return;
+            }
+
+            let cur_chunk = self.current_chunk_footer.get();
+
+            // Deallocate all chunks except the current one
+            let prev_chunk = cur_chunk.as_ref().prev.replace(EMPTY_CHUNK.get());
+            dealloc_chunk_list(prev_chunk);
+
+            // Reset the bump finger to the end of the chunk.
+            debug_assert!(
+                is_pointer_aligned_to(cur_chunk.as_ptr(), MIN_ALIGN),
+                "bump pointer {cur_chunk:#p} should be aligned to the minimum alignment of {MIN_ALIGN:#x}"
+            );
+            cur_chunk.as_ref().ptr.set(cur_chunk.cast());
+
+            debug_assert!(
+                self.current_chunk_footer.get().as_ref().prev.get().as_ref().is_empty(),
+                "We should only have a single chunk"
+            );
+            debug_assert_eq!(
+                self.current_chunk_footer.get().as_ref().ptr.get(),
+                self.current_chunk_footer.get().cast(),
+                "Our chunk's bump finger should be reset to the start of its allocation"
+            );
+        }
+    }
+}
+
+/// `Drop` implementation for `Arena`.
+impl<const MIN_ALIGN: usize> Drop for Arena<MIN_ALIGN> {
+    fn drop(&mut self) {
+        unsafe {
+            dealloc_chunk_list(self.current_chunk_footer.get());
+        }
+    }
+}
+
+#[inline]
+unsafe fn dealloc_chunk_list(mut footer: NonNull<ChunkFooter>) {
+    unsafe {
+        while !footer.as_ref().is_empty() {
+            let f = footer;
+            footer = f.as_ref().prev.get();
+            alloc::dealloc(f.as_ref().data.as_ptr(), f.as_ref().layout);
+        }
+    }
+}
